Dojin p

+++ src/main/java/idv/wantedpreonboardingbackend/jobopening/repository/JobopeningRepository.java
@@ -3,6 +3,6 @@ package idv.wantedpreonboardingbackend.jobopening.repository;
 import idv.wantedpreonboardingbackend.jobopening.domain.JobopeningEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobopeningRepository extends JpaRepository<JobopeningEntity, Integer> {
+public interface JobopeningRepository extends JpaRepository<JobopeningEntity, Long> {
 
 }
